requirements.txt
langchain
langchain-community
langchain-ollama
streamlit
requests
python-dotenv
streamlit-extras
streamlit-echarts
matplotlib
